# lib/api-zod/src/generated/types/serviceUpdate.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Local Service Finder API
 * OpenAPI spec version: 0.1.0
 */

export interface ServiceUpdate {
  title?: string;
  category?: string;
  description?: string;
  price?: number;
  location?: string;
}